Read access levels from user's LDAP entry.
During authentication (i.e. login), read the access levels information
from the LDAP entry. It's possible to store both global access level and
per-project access levels. The default objectClass is mantisUser, and
the attribute that stores the access levels is mantisAccessLevel. The
name of the objectClass is actually being used.

For global access level, the attribute has the format:
    mantisAccessLevel: <access_level>
For per-project levels, the attribute has the format
    mantisAccessLevel: <project_name>:<access_level>

For example:
    objectClass: mantisUser
    mantisAccessLevel: 10
    mantisAccessLevel: project 1:40
    mantisAccessLevel: project 2:55
The user would get global access level of 10, 40 for "project 1", and 55
for "project 2".

Note that changing the access levels via the web interface has no
effect.